+++ HolidaySales/docs/changelog.md
@@ -3,6 +3,7 @@
 
 ### Version 0.2.0
 * Update for Stardew 1.6
+* Now favors using the location context ID over parsing the map name, if that is available.
 
 ### Version 0.1.1
 * Mod-added areas that do not follow convention are now considered their own region.